requirements.txt
httpx
fastapi
slowapi
uvicorn
python-dotenv